+++ packages/swarm-components/lib/Button.js
@@ -1,7 +1,6 @@
 function _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }
 
 import * as React from 'react';
-import cx from 'classnames';
 import Icon from './Icon';
 
 /**
@@ -74,3 +73,101 @@ const Button = props => {
 };
 
 export default Button;
+Button.__docgenInfo = {
+  "description": "",
+  "methods": [],
+  "displayName": "Button",
+  "props": {
+    "bordered": {
+      "required": false,
+      "flowType": {
+        "name": "boolean"
+      },
+      "description": "The bordered button"
+    },
+    "className": {
+      "required": false,
+      "flowType": {
+        "name": "string"
+      },
+      "description": "Classes for additional styles to be applied"
+    },
+    "disabled": {
+      "required": false,
+      "flowType": {
+        "name": "boolean"
+      },
+      "description": "Indicates whether the button is disabled or not"
+    },
+    "icon": {
+      "required": false,
+      "flowType": {
+        "name": "string"
+      },
+      "description": "Supports icon classes found at https://meetup.github.io/swarm-icons/"
+    },
+    "inverted": {
+      "required": false,
+      "flowType": {
+        "name": "boolean"
+      },
+      "description": "Inverted styles for dark backgrounds"
+    },
+    "neutral": {
+      "required": false,
+      "flowType": {
+        "name": "boolean"
+      },
+      "description": "The neutral style"
+    },
+    "onClick": {
+      "required": false,
+      "flowType": {
+        "name": "signature",
+        "type": "function",
+        "raw": "(*) => void",
+        "signature": {
+          "arguments": [{
+            "name": "",
+            "type": {
+              "name": "unknown"
+            }
+          }],
+          "return": {
+            "name": "void"
+          }
+        }
+      },
+      "description": "The function invoked when interacting with Button"
+    },
+    "primary": {
+      "required": false,
+      "flowType": {
+        "name": "boolean"
+      },
+      "description": "The primary style"
+    },
+    "right": {
+      "required": false,
+      "flowType": {
+        "name": "boolean"
+      },
+      "description": "Aligns the icon to the right"
+    },
+    "small": {
+      "required": false,
+      "flowType": {
+        "name": "boolean"
+      },
+      "description": "Use the small button size"
+    },
+    "children": {
+      "required": true,
+      "flowType": {
+        "name": "ReactNode",
+        "raw": "React.Node"
+      },
+      "description": ""
+    }
+  }
+};