+++ packages/swarm-docs/gatsby-config.js
@@ -4,11 +4,19 @@ module.exports = {
   },
   plugins: [
     'gatsby-plugin-react-helmet',
+    'gatsby-plugin-flow',
     {
       resolve: `gatsby-source-filesystem`,
       options: {
         name: `images`,
         path: `${__dirname}/src/images`,
+      },
+    },
+    {
+      resolve: `gatsby-source-filesystem`,
+      options: {
+        name: 'pages',
+        path: `${__dirname}/src/pages/`,
       },
     },
     'gatsby-transformer-sharp',
@@ -34,4 +42,4 @@ module.exports = {
     },
     'gatsby-plugin-offline',
   ],
-}
+};