+++ packages/swarm-docs/src/components/layout-mdx.js
@@ -19,11 +19,18 @@ const PreComponent = ({ className, ...props }) => {
   )
 }
 
+// apply swarm classnames to classes used in markdown
+const H1 = props => <h1 className="text--display" {...props}/>
+const H2 = props => <h2 className="text--pageTitle" {...props}/>
+const H3 = props => <h3 className="text--big" {...props}/>
+const P = props => <p className="text--body" {...props}/>
+const A = props => <a className="text--link" {...props}/>
+
 export default class MDXLayout extends React.Component {
   render() {
     return (
       <Layout>
-        <MDXProvider components={{ pre: PreComponent }}>
+        <MDXProvider components={{ pre: PreComponent, h1: H1, h2: H2, h3: H3, p: P, a: A }}>
           {this.props.children}
         </MDXProvider>
       </Layout>