+++ packages/swarm-constants/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-constants@0.2.0...@meetup/swarm-constants@0.2.2) (2019-02-21)
+
+
+### Bug Fixes
+
+* updating constants dependency ([743966f](https://github.com/meetup/swarm-ui/commit/743966f))
+
+
+
+
+
+# [0.2.0](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-constants@0.1.1...@meetup/swarm-constants@0.2.0) (2019-02-19)
+
+
+### Features
+
+* removing meetup-web-components dependency ([ac8e398](https://github.com/meetup/swarm-ui/commit/ac8e398))
+
+
+
+
+
 ## [0.1.1](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-constants@0.1.0...@meetup/swarm-constants@0.1.1) (2019-02-07)
 
 **Note:** Version bump only for package @meetup/swarm-constants