+++ packages/swarm-styles/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-styles@0.1.2...@meetup/swarm-styles@0.1.4) (2019-02-21)
+
+
+### Bug Fixes
+
+* updating constants dependency ([743966f](https://github.com/meetup/swarm-ui/commit/743966f))
+
+
+
+
+
 ## [0.1.2](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-styles@0.1.1...@meetup/swarm-styles@0.1.2) (2019-02-11)
 
 **Note:** Version bump only for package @meetup/swarm-styles